requirements_cloud.txt
streamlit>=1.25.0
python-dotenv>=1.0.0
langchain>=0.1.230
langchain-community>=0.0.20
langchain-pinecone>=0.1.0
sentence-transformers>=2.2.2
pinecone-client>=3.0.0
groq>=0.4.0
requests>=2.32.0
pypdf>=3.0.0
langchain-text-splitters>=0.0.1
